+++ prisma/migrations/20251216022842_add_mobile_payment_request_id/migration.sql
@@ -1,0 +1,5 @@
+-- AlterTable
+ALTER TABLE "orders" ADD COLUMN     "mobilePaymentRequestId" TEXT;
+
+-- CreateIndex
+CREATE INDEX "orders_mobilePaymentRequestId_idx" ON "orders"("mobilePaymentRequestId");